Fix story viewer race condition bug
StoryViewer feeds PeerStoriesView with data through the pager, but the opposite is also true in PeerStoriesView.updatePosition. This double update can create issues when a story viewer is opened twice with different parameters. It is possible that PeerStoriesView.updatePosition is called before PeerStoriesView receives the new data, and in this case it overrides the new data from the ancestor and the old story is shown.

The least invasive solution is to clear the pager state when the viewer is closed (peer views are cached anyway!), so that state will be clear on the next open (thanks to peer.reset).